+++ backend/core/app/src/main/java/com/mindbridge/core/domains/postReaction/PostReactionService.java
@@ -4,9 +4,12 @@ import com.mindbridge.core.domains.post.PostService;
 import com.mindbridge.core.domains.postReaction.dto.ReceivedPostReactionDto;
 import com.mindbridge.core.domains.postReaction.dto.ResponsePostReactionDto;
 import com.mindbridge.data.domains.postReaction.PostReactionRepository;
+import com.mindbridge.data.domains.user.UserRepository;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
+import org.springframework.web.client.HttpClientErrorException;
 
 import java.beans.Transient;
 import java.util.Optional;
@@ -22,6 +25,9 @@ public class PostReactionService {
 	@Autowired
 	private PostService postService;
 
+	@Autowired
+	private UserRepository userRepository;
+
 	public PostReactionService(PostReactionRepository postReactionRepository) {
 		this.postReactionRepository = postReactionRepository;
 	}
@@ -31,31 +37,33 @@ public class PostReactionService {
 	}
 
 	public Optional<ResponsePostReactionDto> setReaction(ReceivedPostReactionDto postReaction) {
-		var reaction = postReactionRepository.getPostReaction(postReaction.getUserId(), postReaction.getPostId());
-		System.out.println(postReaction);
-		if (reaction.isPresent()) {
-			var react = reaction.get();
-			if (react.getLiked() == postReaction.getLiked()) {
-				postReactionRepository.deleteById(react.getId());
-				return Optional.empty();
-			}
-			else {
-				react.setLiked(postReaction.getLiked());
-				var result = postReactionRepository.save(react);
-				var post = postService.getPostById(result.getPost().getId());
-				return Optional.of(ResponsePostReactionDto.builder().id(result.getId()).liked(result.getLiked())
+		var currentUser = userRepository.findById(postReaction.getUserId()).orElseThrow();
+		if (currentUser.isEmailVerified()) {
+			var reaction = postReactionRepository.getPostReaction(postReaction.getUserId(), postReaction.getPostId());
+			System.out.println(postReaction);
+			if (reaction.isPresent()) {
+				var react = reaction.get();
+				if (react.getLiked() == postReaction.getLiked()) {
+					postReactionRepository.deleteById(react.getId());
+					return Optional.empty();
+				} else {
+					react.setLiked(postReaction.getLiked());
+					var result = postReactionRepository.save(react);
+					var post = postService.getPostById(result.getPost().getId());
+					return Optional.of(ResponsePostReactionDto.builder().id(result.getId()).liked(result.getLiked())
 						.userId(result.getAuthor().getId()).postId(post.getId()).authorId(post.getAuthor().getId())
 						.isFirstReaction(false).build());
-			}
-		}
-		else {
-			var postReact = PostReactionMapper.MAPPER.dtoToPostReaction(postReaction);
-			var result = postReactionRepository.save(postReact);
-			var post = postService.getPostById(result.getPost().getId());
-			return Optional.of(ResponsePostReactionDto.builder().id(result.getId()).liked(result.getLiked())
+				}
+			} else {
+				var postReact = PostReactionMapper.MAPPER.dtoToPostReaction(postReaction);
+				var result = postReactionRepository.save(postReact);
+				var post = postService.getPostById(result.getPost().getId());
+				return Optional.of(ResponsePostReactionDto.builder().id(result.getId()).liked(result.getLiked())
 					.userId(result.getAuthor().getId()).postId(post.getId()).authorId(post.getAuthor().getId())
 					.isFirstReaction(true).build());
+			}
 		}
+		throw new HttpClientErrorException(HttpStatus.FORBIDDEN);
 	}
 
 }
